+++ client/web/src/enterprise/insights/components/insights-view-grid/components/backend-insight/BackendInsight.tsx
@@ -23,6 +23,7 @@ import {
     DrillDownInsightCreationFormValues,
     BackendInsightChart,
 } from './components'
+import { useSeriesToggle } from './components/backend-insight-chart/use-series-toggle'
 import { parseSeriesDisplayOptions } from './components/drill-down-filters-panel/drill-down-filters/utils'
 
 import styles from './BackendInsight.module.scss'
@@ -44,6 +45,7 @@ export const BackendInsightView: React.FunctionComponent<React.PropsWithChildren
 
     const { currentDashboard, dashboards } = useContext(InsightContext)
     const { getBackendInsightData, createInsight, updateInsight } = useContext(CodeInsightsBackendContext)
+    const { toggle, isSeriesSelected, isSeriesHovered, setHoveredId } = useSeriesToggle()
 
     // Visual line chart settings
     const [zeroYAxisMin, setZeroYAxisMin] = useState(false)
@@ -184,7 +186,15 @@ export const BackendInsightView: React.FunctionComponent<React.PropsWithChildren
             ) : state.status === LazyQueryStatus.Error ? (
                 <BackendInsightErrorAlert error={state.error} />
             ) : (
-                <BackendInsightChart {...state.data} locked={insight.isFrozen} onDatumClick={trackDatumClicks} />
+                <BackendInsightChart
+                    {...state.data}
+                    locked={insight.isFrozen}
+                    onDatumClick={trackDatumClicks}
+                    toggle={toggle}
+                    isSeriesSelected={isSeriesSelected}
+                    isSeriesHovered={isSeriesHovered}
+                    setHoveredId={setHoveredId}
+                />
             )}
             {
                 // Passing children props explicitly to render any top-level content like